+++ lib/ts/recipe/thirdparty/providers/gitlab.ts
@@ -13,6 +13,8 @@
  * under the License.
  */
 
+import NormalisedURLDomain from "../../../normalisedURLDomain";
+import NormalisedURLPath from "../../../normalisedURLPath";
 import { TypeProvider, ProviderInput } from "../types";
 import NewProvider from "./custom";
 // import NormalisedURLDomain from "../../../normalisedURLDomain";
@@ -20,10 +22,6 @@ import NewProvider from "./custom";
 export default function Gitlab(input: ProviderInput): TypeProvider {
     if (input.config.name === undefined) {
         input.config.name = "Gitlab";
-    }
-
-    if (input.config.oidcDiscoveryEndpoint === undefined) {
-        input.config.oidcDiscoveryEndpoint = "https://gitlab.com";
     }
 
     const oOverride = input.override;
@@ -39,9 +37,12 @@ export default function Gitlab(input: ProviderInput): TypeProvider {
 
             if (config.oidcDiscoveryEndpoint === undefined) {
                 if (config.additionalConfig !== undefined && config.additionalConfig.gitlabBaseUrl !== undefined) {
-                    config.oidcDiscoveryEndpoint = config.additionalConfig.gitlabBaseUrl;
+                    const oidcDomain = new NormalisedURLDomain(config.additionalConfig.gitlabBaseUrl);
+                    const oidcPath = new NormalisedURLPath("/.well-known/openid-configuration");
+
+                    config.oidcDiscoveryEndpoint = oidcDomain.getAsStringDangerous() + oidcPath.getAsStringDangerous();
                 } else {
-                    config.oidcDiscoveryEndpoint = "https://gitlab.com";
+                    config.oidcDiscoveryEndpoint = "https://gitlab.com/.well-known/openid-configuration";
                 }
             }